# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Portfolio Service API",
    "description": "Declarative Zod Validated & Programmatically Generated API Spec",
    "version": "1.0.0"
  },
  "paths": {
    "/api/case-studies": {
      "get": {
        "summary": "Retrieve published case studies",
        "description": "Fetches a list of published case studies including title, slug, primary language, and tags.",
        "responses": {
          "200": {
            "description": "Successful retrieval of case studies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseStudySummary"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit draft technical post-mortem case study",
        "description": "Submits a new technical case study or prototype post-mortem as an unpublished draft record.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseStudySubmission"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Case study submitted successfully in unpublished draft state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseStudyPostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed payload fields or duplicate slug",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Administrator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/projects/{slug}/image": {
      "post": {
        "summary": "Upload and link project hero image",
        "description": "Stores a validated media asset and updates the target case study hero image URL for authorized administrators.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file buffer (PNG, JPEG, WebP, GIF, SVG, AVIF; max 5MB)"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image asset uploaded and persisted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "hero_image_url": {
                          "type": "string"
                        },
                        "key": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation or payload error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Administrator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear project hero image",
        "description": "Clears the hero image URL associated with a project for authorized administrators.",
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hero image cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "hero_image_url": {
                          "type": "null"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Administrator access required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ClerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token or administrative JWT"
      },
      "CronSecretAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Secret",
        "description": "Cron secret bearer authorization header"
      }
    },
    "schemas": {
...
      "CaseStudySummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "slug": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "primary_language": {
            "type": "string",
            "minLength": 1
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "primary_language",
          "tags"
        ],
        "additionalProperties": false
      },
      "CaseStudySubmission": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9-_]+$"
          },
          "primary_language": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "editorial_content": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "summary_markdown": {
            "type": "string"
          },
          "architectural_narrative": {
            "type": "string"
          },
          "narrative": {
            "type": "string"
          },
          "architectural_narrative_html": {
            "type": "string"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "minItems": 1,
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1
                }
              }
            ]
          },
          "github_url": {
            "type": "string"
          },
          "hero_image_url": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri",
            "nullable": true
          },
          "heroImageUrl": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri",
            "nullable": true
          }
        },
        "required": [
          "title",
          "slug",
          "tags"
        ],
        "additionalProperties": false
      },
      "CaseStudyPostResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "primary_language": {
                "type": "string"
              },
              "editorial_content": {
                "type": "string"
              },
              "architectural_narrative": {
                "type": "string"
              },
              "published": {
                "type": "boolean"
              },
              "tags": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "slug",
              "title",
              "primary_language",
              "editorial_content",
              "architectural_narrative",
              "published",
              "tags"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "success",
          "data"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}